small refactor - moved source code to src directory - moved build output to dist directory - broke bundle to separate app and vendor bundles - renamed the main component to Hello - updated gitignore - added an editor config file
