--- app.js
@@ -1,5 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import HMU from "./module.js";
-
-ReactDOM.render(<HMU name={["HMU", "GoDaddy"]} special={true} />, document.getElementById("app"));
