+++ index.html
@@ -1,10 +1,10 @@
+<!doctype html>
 <html>
-  <head>
-  </head>
-  <body>
-    <div id="app">
-
-    </div>
-    <script src="bundle.js" type="text/javascript"></script>
-  </body>
+<head>
+</head>
+<body>
+<div id="app"></div>
+<script src="vendor.js" type="text/javascript"></script>
+<script src="app.js" type="text/javascript"></script>
+</body>
 </html>
